refactor: move paging from services to lib
