+++ src/actions/news/read.ts
@@ -3,7 +3,7 @@ import { safeServerCall } from '@/actions/safeServerCall'
 import { readNews, readNewsCurrent, readOldNewsPage } from '@/services/news/read'
 import type { ExpandedNewsArticle, NewsCursor, SimpleNewsArticle } from '@/services/news/Types'
 import type { ActionReturn } from '@/actions/Types'
-import type { ReadPageInput } from '@/services/paging/Types'
+import type { ReadPageInput } from '@/lib/paging/Types'
 
 export async function readOldNewsPageAction<const PageSize extends number>(
     readPageImput: ReadPageInput<PageSize, NewsCursor>
