+++ src/contexts/paging/EventArchivePaging.tsx
@@ -2,7 +2,7 @@
 import generatePagingProvider, { generatePagingContext } from './PagingGenerator'
 import { readArchivedEventsPage } from '@/actions/events/read'
 import type { EventArchiveCursor, EventArchiveDetails, EventExpanded } from '@/services/events/Types'
-import type { ReadPageInput } from '@/services/paging/Types'
+import type { ReadPageInput } from '@/lib/paging/Types'
 
 export type PageSizeEventArchive = 12
 const fetcher = async (x: ReadPageInput<PageSizeEventArchive, EventArchiveCursor, EventArchiveDetails>) => {
