+++ src/contexts/paging/ImageCollectionPaging.tsx
@@ -1,7 +1,7 @@
 'use client'
 import generatePagingProvider, { generatePagingContext } from './PagingGenerator'
 import { readImageCollectionsPageAction } from '@/actions/images/collections/read'
-import type { ReadPageInput } from '@/services/paging/Types'
+import type { ReadPageInput } from '@/lib/paging/Types'
 import type { ImageCollectionCursor, ImageCollectionPageReturn } from '@/services/images/collections/Types'
 
 export type PageSizeImageCollection = 12
