+++ src/contexts/paging/ImagePaging.tsx
@@ -1,7 +1,7 @@
 'use client'
 import generatePagingProvider, { generatePagingContext } from './PagingGenerator'
 import { readImagesPageAction } from '@/actions/images/read'
-import type { ReadPageInput } from '@/services/paging/Types'
+import type { ReadPageInput } from '@/lib/paging/Types'
 import type { Image } from '@prisma/client'
 import type { ImageCursor, ImageDetails } from '@/services/images/Types'
