+++ src/contexts/paging/JobAdInactivePaging.tsx
@@ -1,7 +1,7 @@
 'use client'
 import generatePagingProvider, { generatePagingContext } from './PagingGenerator'
 import { readInactiveJobAdsPageAction } from '@/actions/career/jobAds/read'
-import type { ReadPageInput } from '@/services/paging/Types'
+import type { ReadPageInput } from '@/lib/paging/Types'
 import type { JobAdInactiveCursor, JobAdInactiveDetails, SimpleJobAd } from '@/services/career/jobAds/Types'
 
 export type PageSizeJobAdInactive = 12
