+++ src/contexts/paging/LockerPaging.tsx
@@ -1,7 +1,7 @@
 'use client'
 import generatePagingProvider, { generatePagingContext } from '@/contexts/paging/PagingGenerator'
 import { readLockerPageAction } from '@/actions/lockers/read'
-import type { ReadPageInput } from '@/services/paging/Types'
+import type { ReadPageInput } from '@/lib/paging/Types'
 import type { LockerCursor, LockerWithReservation } from '@/services/lockers/Types'
 
 export type PageSizeLocker = 20
