+++ src/contexts/paging/OldNewsPaging.tsx
@@ -1,7 +1,7 @@
 'use client'
 import generatePagingProvider, { generatePagingContext } from './PagingGenerator'
 import { readOldNewsPageAction } from '@/actions/news/read'
-import type { ReadPageInput } from '@/services/paging/Types'
+import type { ReadPageInput } from '@/lib/paging/Types'
 import type { NewsCursor, SimpleNewsArticle } from '@/services/news/Types'
 
 export type PageSizeOldNews = 20
