+++ src/contexts/paging/SchoolPaging.tsx
@@ -1,7 +1,7 @@
 'use client'
 import generatePagingProvider, { generatePagingContext } from './PagingGenerator'
 import { readSchoolsPageAction } from '@/education/schools/read'
-import type { ReadPageInput } from '@/services/paging/Types'
+import type { ReadPageInput } from '@/lib/paging/Types'
 import type { ExpandedSchool, SchoolCursor } from '@/education/schools/Types'
 
 export type PageSizeSchool = 8
