+++ src/services/career/jobAds/read.ts
@@ -3,9 +3,9 @@ import { jobAdArticleRealtionsIncluder, simpleJobAdArticleRealtionsIncluder } fr
 import { CompanyRelationIncluder } from '@/career/companies/ConfigVars'
 import { ServerError } from '@/services/error'
 import { ServiceMethodHandler } from '@/services/ServiceMethodHandler'
-import { cursorPageingSelection } from '@/services/paging/cursorPageingSelection'
+import { cursorPageingSelection } from '@/lib/paging/cursorPageingSelection'
 import type { ExpandedJobAd, JobAdInactiveCursor, JobAdInactiveDetails, SimpleJobAd } from './Types'
-import type { ReadPageInput } from '@/services/paging/Types'
+import type { ReadPageInput } from '@/lib/paging/Types'
 
 /**
  * This handler reads a jobAd by id or articleName and order
