+++ src/services/events/read.ts
@@ -1,9 +1,9 @@
 import 'server-only'
 import { eventFilterSeletion } from './ConfigVars'
 import { ServiceMethodHandler } from '@/services/ServiceMethodHandler'
-import { cursorPageingSelection } from '@/services/paging/cursorPageingSelection'
+import { cursorPageingSelection } from '@/lib/paging/cursorPageingSelection'
 import { getOsloTime } from '@/lib/dates/getOsloTime'
-import type { ReadPageInput } from '@/services/paging/Types'
+import type { ReadPageInput } from '@/lib/paging/Types'
 import type { EventArchiveCursor, EventArchiveDetails } from './Types'
 
 export const read = ServiceMethodHandler({
