+++ src/services/omegaquotes/read.ts
@@ -1,9 +1,9 @@
 import 'server-only'
 import { omegaQuoteFilterSelection } from './CofigVars'
-import { cursorPageingSelection } from '@/services/paging/cursorPageingSelection'
+import { cursorPageingSelection } from '@/lib/paging/cursorPageingSelection'
 import { prismaCall } from '@/services/prismaCall'
 import prisma from '@/prisma'
-import type { ReadPageInput } from '@/services/paging/Types'
+import type { ReadPageInput } from '@/lib/paging/Types'
 import type { OmegaquoteCursor, OmegaquoteFiltered } from '@/services/omegaquotes/Types'
 
 export async function readQuotesPage<const PageSize extends number>(
